[android] Fixed review notes

+++ android/src/com/mapswithme/util/FileUploadWorker.java
@@ -6,6 +6,8 @@ import androidx.annotation.NonNull;
 import androidx.work.Worker;
 import androidx.work.WorkerParameters;
 import com.google.gson.Gson;
+import com.mapswithme.util.log.Logger;
+import com.mapswithme.util.log.LoggerFactory;
 
 import java.io.File;
 import java.util.Objects;
@@ -13,6 +15,8 @@ import java.util.Objects;
 public class FileUploadWorker extends Worker
 {
   static final String PARAMS = "params";
+  private static final Logger LOGGER = LoggerFactory.INSTANCE.getLogger(LoggerFactory.Type.NETWORK);
+  private static final String TAG = FileUploadWorker.class.getSimpleName();
 
   public FileUploadWorker(@NonNull Context context, @NonNull WorkerParameters workerParams)
   {
@@ -27,14 +31,21 @@ public class FileUploadWorker extends Worker
 
     Gson gson = new Gson();
     HttpPayload payload = gson.fromJson(rawJson, HttpPayload.class);
+    LOGGER.d(TAG, "Payload '" + payload + "' going to be uploaded");
+
     File file = new File(payload.getFilePath());
     if (!file.exists())
-      return Result.success();
+      return Result.failure();
 
     HttpUploader uploader = new HttpUploader(payload);
     HttpUploader.Result result = uploader.upload();
+    LOGGER.d(TAG, "Upload finished with result '" + result + "' ");
 
-    file.delete();
-    return result.getHttpCode() / 100 != 2 ? Result.failure() : Result.success();
+    if (result.getHttpCode() / 100 != 2)
+      return Result.retry();
+
+    boolean isDeleted = file.delete();
+    LOGGER.d(TAG, "File deleted: " + isDeleted + " " + payload.getFilePath());
+    return Result.success();
   }
 }
